Rename files to not have unnecessary tt_smi_ prefix
- No functional changes
- the "tt_smi_" prefix on filenames was making the files inconsistent
  and was not providing any function
- Updated all references and tests
