+++ tt_smi/utils.py
